+++ Core/MIPS/IR/IRInterpreter.h
@@ -11,9 +11,6 @@ u32 IRRunBreakpoint(u32 pc);
 u32 IRRunMemCheck(u32 pc, u32 addr);
 u32 IRInterpret(MIPSState *ms, const IRInst *inst);
 
-void IRApplyRounding();
-void IRRestoreRounding();
-
 template <uint32_t alignment>
 u32 RunValidateAddress(u32 pc, u32 addr, u32 isWrite) {
 	static_assert(alignment <= 16);
